First part of the Windows port: fixed compilation errors and warnings

+++ src/display-tracking/main.cpp
@@ -22,6 +22,7 @@
 #include "utils/points.hpp"
 #include <iostream>
 #include <opencv2/highgui/highgui.hpp>
+#include <opencv2/imgproc.hpp>
 
 void
 print_usage()
@@ -98,7 +99,7 @@ run_program(int argc, char **argv)
     cv::imwrite(output_pathname->c_str(), img);
   } else {
     cv::imshow("Image", img);
-    cv::waitKey(1000*wait_time);
+    cv::waitKey((int) (1000*wait_time));
   }
   
   return 0;
